[TE] detection - input data fetcher interface
