+++ thirdeye/thirdeye-pinot/src/main/java/com/linkedin/thirdeye/detection/InputDataFetcher.java
@@ -16,64 +16,17 @@
 
 package com.linkedin.thirdeye.detection;
 
-import com.google.common.collect.Multimap;
-import com.linkedin.thirdeye.dataframe.DataFrame;
-import com.linkedin.thirdeye.dataframe.util.MetricSlice;
-import com.linkedin.thirdeye.datalayer.dto.DatasetConfigDTO;
-import com.linkedin.thirdeye.datalayer.dto.EventDTO;
-import com.linkedin.thirdeye.datalayer.dto.MergedAnomalyResultDTO;
-import com.linkedin.thirdeye.datalayer.dto.MetricConfigDTO;
-import com.linkedin.thirdeye.detection.spi.model.AnomalySlice;
-import com.linkedin.thirdeye.detection.spi.model.EventSlice;
 import com.linkedin.thirdeye.detection.spi.model.InputData;
 import com.linkedin.thirdeye.detection.spi.model.InputDataSpec;
-import java.util.Collections;
-import java.util.HashMap;
-import java.util.Map;
-
 
 /**
- * Input data fetcher.
+ * Input data fetcher interface.
  * For components to fetch the input data it need.
  */
-public class InputDataFetcher {
-  private DataProvider provider;
-  private long configId;
 
-  public InputDataFetcher(DataProvider provider, long configId) {
-    this.provider = provider;
-    this.configId = configId;
-  }
-
+public interface InputDataFetcher {
   /**
-   * Fetch data for input data spec
+   * fetch data for input data spec
    */
-  public InputData fetchData(InputDataSpec inputDataSpec) {
-    Map<MetricSlice, DataFrame> timeseries = provider.fetchTimeseries(inputDataSpec.getTimeseriesSlices());
-    Map<MetricSlice, DataFrame> aggregates =
-        provider.fetchAggregates(inputDataSpec.getAggregateSlices(), Collections.<String>emptyList());
-    Multimap<AnomalySlice, MergedAnomalyResultDTO> existingAnomalies =
-        provider.fetchAnomalies(inputDataSpec.getAnomalySlices(), configId);
-    Multimap<EventSlice, EventDTO> events = provider.fetchEvents(inputDataSpec.getEventSlices());
-    Map<Long, MetricConfigDTO> metrics = provider.fetchMetrics(inputDataSpec.getMetricIds());
-    Map<String, DatasetConfigDTO> datasets = provider.fetchDatasets(inputDataSpec.getDatasetNames());
-
-    Map<Long, DatasetConfigDTO> datasetForMetricId = fetchDatasetForMetricId(provider, inputDataSpec);
-    return new InputData(inputDataSpec, timeseries, aggregates, existingAnomalies, events, metrics, datasets, datasetForMetricId);
-  }
-
-  private Map<Long, DatasetConfigDTO> fetchDatasetForMetricId(DataProvider provider, InputDataSpec inputDataSpec) {
-    Map<Long, MetricConfigDTO> metrics = provider.fetchMetrics(inputDataSpec.getMetricIdsForDatasets());
-    Map<Long, String> metricIdToDataSet = new HashMap<>();
-    for (Map.Entry<Long, MetricConfigDTO> entry : metrics.entrySet()){
-      metricIdToDataSet.put(entry.getKey(), entry.getValue().getDataset());
-    }
-    Map<String, DatasetConfigDTO> datasets = provider.fetchDatasets(metricIdToDataSet.values());
-    Map<Long, DatasetConfigDTO> result = new HashMap<>();
-    for (Map.Entry<Long, MetricConfigDTO> entry : metrics.entrySet()){
-      result.put(entry.getKey(), datasets.get(entry.getValue().getDataset()));
-    }
-    return result;
-  }
-
+  InputData fetchData(InputDataSpec inputDataSpec);
 }
